LegacyOpenerTest: fix invalid macro
The testPaulMacro method relied on an undocumented and soon-to-be
invalid behavior: an output variable named "nResults" -- i.e.,
matching a built-in IJ1 macro function. This dirty hack worked
due to an implementation detail of how outputs used to get populated.
But that detail will change in the next commit, so we must change the
test or else it will start failing (correctly so!).

If you want to save the results of a built-in function like nResults,
the correct approach is to assign it to an actual output variable
whose name does not clash with any built-in macro language keywords.