LegacyService: fix shift+click to open scripts
It used to be that you could shift-click a script in the menu structure,
and it would open in the Script Editor rather than executing it.

Unfortunately, I recently rewrote the script discovery mechanism, which
broke this feature; see:

https://github.com/scijava/scijava-common/commit/ff73f88a96f5d06a8565adfb633c3fe1ce8ef218#diff-b7a8ea280aed4f4f6b78b86f6129b3beR195

The issue is that the getPath() method of ScriptInfo no longer returns a
fully valid path to the file on disk, since all scripts are now treated
as URLs internally. This paved the way for script discovery inside JAR
files, but broke an assumption here that scripts are always files.

The fix is to read the script via the ScriptInfo#getReader() method,
when a BufferedReader is available. In theory, this makes shift-click
work for scripts stored inside JAR files too, although I did not test.

For more details and discussion, see:
http://forum.imagej.net/t/shift-clicking-on-menu-item-to-open-script-editor/2617

Thanks to Jan Eglinger for reporting this problem.