IJ1Helper: fix the isRMIEnabled method
It was calling out to the old ij.OtherInstance class, which uses
an obsolete and now erroneous mechanism for determine whether the
"Run single instance listener" option is currently set.

This was the _only_ use of ij.OtherInstance. Rather than worrying
about that class, let's just ask ij.Prefs directly what the value is.
(Maybe there was a reason we did not do it this way originally, but
that knowledge is lost to me -- so we try the easy way first.)

This fix should make it possible to launch multiple instances of
ImageJ on Windows systems again.