LegacyService: fix shift+click script open code
Now, when shift+clicking a menu item that happens to be a script,
we try harder to discern the location of that script on disk.

This is now possible because scripts now know themselves as URLs,
and we can check whether a URL exists as a file on disk.

This also updates the logic for the non-file-on-disk case to call
setFileName(File) instead of setFileName(String), which should now
set the script language properly.

See:
http://forum.imagej.net/t/shift-clicking-on-menu-item-to-open-script-editor/2617