IJ1Helper: make saveDialog smarter about extension
When given a file with a name and extension in the "File file"
parameter, but null for the "String extensions", the file chooser does
not preserve the file extension. Consequently, for use cases like the
Script Editor saving a file for the first time, the extension was
truncated. (And later, for reasons I did not investigate, the final
character of the extensionless filename was also cut off.)

This change now extracts and uses the file extension from the
"File file" parameter in the case where it makes sense to do so.

See also imagej/imagej#157.