Fix duplicate entries in image selections
Previously, image selections (e.g. in a Swing UI) would show three
entries in a dropdown for each image opened as a Dataset. The issue was
that multiple converters would add convertible inputs to the set of
compatible inputs without checking if another converter has added a
different representation of the same data (e.g. ImageTitle,
ImageDisplay, Dataset).

This commit changes the implementation in ImageTitleToImagePlusConverter
to check if a Dataset is available, i.e. it preferes Datasets. Similar
changes check for an available Dataset in the
ImageDisplayToImagePlusConverter to prefer that.