Exclude FocusSearchBar from search results
This is a little hacky. I was going to put "visible = false" instead,
but if you do that, the machinery skips adding it to the ImageJ 1.x menu
structure, and then the keyboard shortcut does not work.

Marking it as "enabled = false" does not actually disable it in the
menus, but does prevent it from appearing in the search results window.