Remove legacy command output tracking machinery
We tried very hard to make ImageJ 1.x commands report generated
and/or manipulated ImagePlus objects as outputs. But the logic for
tracking it was complex and error-prone. Ultimately, there is no
need for this sort of output tracking, and it create all kinds of
weirdness. Some commands would spontaneously generate tables. Some
commands (maybe all commands?) would have astonishing side effects
like the active threshold being cleared before execution.

So we purge all this convoluted machinery, in favor of a dirt simple
LegacyCommand that just runs the command and nothing else.