Add a utility class for working with macros
The first use case is for ImageJ-OMERO, which wants
to run macros headless, with ImagePlus inputs.

Suppose we have the following macro:

    #@ ImagePlus imp
    #@output ImagePlus output
    run("Duplicate...", " ");
    output = getImageID();

Now suppose we execute this macro from the CLI as follows:

    ImageJ --ij2 --headless --run duplicator.ijm

It will do nothing. Firstly, imp input cannot be populated because there
is no active image in a freshly started ImageJ. (In ImageJ-OMERO's case,
the input image will be fed in to the execution externally.) Let's work
around this issue for testing purposes by creating a second macro:

    open("/path/to/myImage.tif");

And then including it in our CLI execution:

    ImageJ --ij2 --headless --run opener.ijm --run duplicator.ijm

Now there will be a populated imp. But there are still problems. In
headless mode, ImagePlus objects do not have ImageWindows, so the
WindowManager is not fully functional. We need a way to set the active
image -- at least using WindowManager.setTempCurrentImage(ImagePlus),
which still works headless -- from a macro. And it would be nice if we
could refer to the ImagePlus by its variable name.

Here comes this commit, offering a new net.imagej.legacy.Macros class
with static utility methods conforming to the ImageJ1 macro language's
call function!

    #@ ImagePlus imp
    #@output ImagePlus output
    print("imp: " + imp);
    call("net.imagej.legacy.Macros.setActiveImage", "imp");
    run("Duplicate...", " ");
    output = getImageID();
    print("output: " + output);

And the output:

    imp: -2
    output: -3
    [INFO] output = img["myImage-1.tif" (-3), 8-bit, 320x200x1x1x1]

There are still plenty of problems with running macros headless,
such as the RoiManager not being available. But this function gives
ImageJ-OMERO a way forward to call headless server-side macros that
operate on images.