Fix the broken ImagePlus-related display viewers
When displaying an ImagePlus, or wrapping an ImageDisplay as ImagePlus
via the legacy layer, we should _not_ create, much less make visible,
a SwingDisplayWindow. The ImagePlus will take care of creating its own
window and showing itself.

So firstly, this commit consolidates the ImagePlusDisplayViewer
interface, AbstractImagePlusDisplayViewer abstract base class, and
DefaultImagePlusDisplayViewer into a single concrete
ImagePlusDisplayViewer class. This is much simpler and easier to
understand. Hopefully nothing downstream needed that extra interface.

Secondly, this commit overrides the view(UserInterface, Display) method
of DisplayViewer to skip the default step of asking the associated
UserInterface to create a display window, in favor of passing null to
view(DisplayWindow, Display). This is fine since we actively do not want
to make a SwingDisplayWindow, but instead want ImagePlus to handle it.

These changes fix a problem where a tiny bogus second window appears
when displaying an ImageDisplay in the legacy user interface.