Rewrite the single instance CLI handling
Instead of overriding the entire ConsoleService, there is now
a ConsoleArgument plugin. This is still slightly hacky (most
ConsoleArgument plugins examine only the leading arguments of
the linked list, removing one or more if appropriate), but
IMHO less so than a full service override. The plugin scans
the entire set of arguments to decide whether and how to
attempt to contact another running instance of ImageJ.

See the javadoc of SingleInstanceArgument for further details.

This commit is dedicated to Christian Tischer.