Move parameter initialization to the end of macros
We now declare global variables with `var` in a prefix of the first line of a macro, and assign their values in a helper function appended to the end of the macro.
To make this possible, we need to fix variable parsing to remove the " (g)" suffix of global variables returned by Interpreter.getVariables().

Fixes #197.