LegacyService: fix bug with LegacyHooks injection
The code was first creating the ImageJ1 instance, and then overriding
the EssentialLegacyHooks with DefaultLegacyHooks. But this resulted in
the LegacyHooks#addMenuItem(String, String) method (and maybe others
that are triggered during ImageJ1 instantiation; I didn't check) not
being overridable in DefaultLegacyHooks.

With this change, the DefaultLegacyHooks are now injected prior to
ImageJ1's instantiation, so that all method hooks are called properly.