Make LegacyOpener return values match contract
When we return a non-null object via a LegacyOpener, we essentially
"claim" the IO process - preventing any other mechanism for attempting
to open the input.

According to the LegacyOpener docs, Boolean.TRUE should be returned if
cancelled, and null if the operation fails. This patch brings behavior
in line with specification.