MacroPreprocessor: resolve Button type inputs
When calling a command from a macro, we usually don't want to show a UI dialog.

If the called command contains inputs of type Button, we resolve these without setting them,
as we assume that buttons and their callbacks provide some interactivity that is not required
when running from macro.

See https://github.com/imagej/imagej-legacy/pull/239 for some discussion.