Add converter image title / ID to Dataset.
This converter converts a String that matches an ImagePlus title or ID to
Dataset. (The ImagePlus must be known to the WindowManager.)

This functionality allows to properly call a SciJava Commands with Dataset
inputs from an IJ1 macros. An IJ1 macro usually refers to an image by its
title or ID.