RoiConversionTest: update expected bounds
In https://github.com/imagej/ImageJ/commit/fa8bee728, ij.gui.Line was
changed to ensure the getBounds() method always returns a bounding
rectangle that contains the given line. This means that bounding boxes
are now typically one pixel wider/taller than they were previously and
requires a corresponding update to our tests.

To understand this change it helps to visualize pixels as squares with
area. Lines between integer points can be thought of as drawn from the
centers of those pixels: so a Line(10, 10, 10, 10) starts at the center
of pixel (10, 10) and ends at the center of pixel (20, 20). Rectangles
are drawn in the spaces "between" the pixels, so a bounding rectangle
that contains this line must start at (9.5, 9.5) and end at (20.5,
20.5) - with a width/height of 11.

If a line started at a half-decimal, e.g. Line(9.5, 9.5, 10, 10) you
could use a smaller bounding box rectangle - (9.5, 9.5) to (19.5, 19.5)
with height/width of 10, in theory, but note that as of
https://github.com/imagej/ImageJ/commit/b1d9081b2b52100fc6e6b4de01e48c051d17b4b2
there is a bug that Lines seem able to start on midpoints but not end on
them.