Adopt the preferred Look & Feel by default
Each Swing-based user interface, including LegacyUI, is responsible for
calling lafService.initLookAndFeel() at the appropriate moment. Any
SwingUI that extends AbstractSwingUI already does this automatically via
the protected createUI() method, but LegacyUI is special and implements
the SwingUI interface directly, due to its diverging requirements from
the other Swing-based UIs.

Note that we must avoid calling initLookAndFeel() in cases where the
application is running headless, because we don't want to trigger
startup of Java' AWT subsystem too eagerly. (I am not certain that calls
to the UIManager trigger such startup, but it wouldn't surprise me, and
it might be different on a per-JVM-implementation basis, so let's
definitely avoid doing it for as long as possible.) As such, this
commit performs the magic when ij1Helper.setVisible(true) is invoked,
just prior to packing and showing the ImageJ application frame.