unregisterImage: actually remove mapping
Previously when unregisterImage was called on an ImagePlus that had a
mapped ImageDisplay, the mapping would be preserved and the display
would be closed, relying on that action to break the mapping.

This created problems when the ImagePlus was *already* being closed,
which is why unregisterImage was being called in the first place. It
would lead to a deadlock on the EDT waiting for a lock to close an
image, but the WindowManager was already locked on the Quit thread that
triggered the unregisterImage call.

Instead we assume unregisterImage is being called for an appropriate
reason: we break the link between ImagePlus and ImageDisplay, and then
close the ImageDisplay.