Add support for so-called "eager" IOPlugins
This is a mechanism by which certain IOPlugins can be labeled with an
"eager" attribute in the @Plugin annotation, and then ImageJ Legacy's
override of File>Open... will give them a chance to handle files before
the original ImageJ goes through its whole pipeline of logic.

This behavior happens even when the "Use SciJava I/O" option is disabled,
so plugins should take care to label themselves "eager" only when they
really *need* to go before standard ImageJ logic. This is currently the
case e.g. for directories, because the original ImageJ always tries to
handle them with Open Image Sequence, but we have cases where we want to
handle directories with particular contents, such as N5 and Zarr, with
plugins instead.