Minimize chance of duplicate convert candidates
the ImageTitle*, ImageDisplay* and Dataset* toImagePlus converters had
the potential to add different views of the same image in their
populateInputCandidates. This is because if, e.g., an ImagePlus is
wrapped to a Dataset (as it is when synchronizing between the layers)
there will be: an image id, display, and dataset created, and each
converter will add its respective instance to the input candidates.

See:
https://forum.image.sc/t/plugin-maintainers-can-you-test-fiji-2-11-0/78852/7

To resolve this potential, we route the populateInputCandidates methods
of each converter through a helper method that checks what is currently
in the candidate list and avoids adding duplicates.

Unfortunately this test is currently done using toString comparison, as
each of ImageTitle, ImageDisplay and Dataset have toString
implementations that essentially are the image title.

Ideally we could explore each Object in the candidates collection and
try to "unwrap" them to their root ImagePlus - however this would be
less extensible and require each converter (or the abstract layer) to
know the input types of ALL the converter implementations, or at least
implement a common interface.