Remove IJ1Helper param from DefaultLegacyHooks
The LegacyService has a setter for its IJ1Helper. Tracking it separately
in the DefaultLegacyHooks creates an opportunity for skew between its
helper and the LegacyService's helper. The LegacyService has an accessor
for its IJ1Helper instance so it is not necessary to cache it in the
DefaultLegacyHooks.

See https://github.com/imagej/imagej-legacy/issues/305