Return value situationally to original ImageJ
In
https://github.com/imagej/imagej-legacy/commit/cc004839c54bf6aa5b4a951f50b953eb4f0383cb
we started returning the path instead of null in
DefaultLegacyOpener#open. This was sensible for File > Open... but broke
File > Open Recent, which depended on a return value of null to pass
through behavior correctly.

To harmonize these scenarios we now return null if the incoming path was
provided (as done in File > Open Recent) but return the actual path if
we displayed a file chooser to the user (as failing to do so would cause
a second chooser to pop).

Closes #314